+++ src/Mammoth.Extensions.DependencyInjection.Tests/TestServices.cs
@@ -24,21 +24,23 @@
 			{
 				if (disposing)
 				{
-					// TODO: dispose managed state (managed objects)
+					// dispose managed state (managed objects)
 				}
 
-				// TODO: free unmanaged resources (unmanaged objects) and override finalizer
-				// TODO: set large fields to null
+				// free unmanaged resources (unmanaged objects) and override finalizer
+				// set large fields to null
 				disposedValue = true;
 			}
 		}
 
-		// // TODO: override finalizer only if 'Dispose(bool disposing)' has code to free unmanaged resources
+		/**
+		// // override finalizer only if 'Dispose(bool disposing)' has code to free unmanaged resources
 		// ~SingletonService()
 		// {
 		//     // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
 		//     Dispose(disposing: false);
 		// }
+		*/
 
 		public void Dispose()
 		{
@@ -68,21 +70,23 @@
 			{
 				if (disposing)
 				{
-					// TODO: dispose managed state (managed objects)
+					// dispose managed state (managed objects)
 				}
 
-				// TODO: free unmanaged resources (unmanaged objects) and override finalizer
-				// TODO: set large fields to null
+				// free unmanaged resources (unmanaged objects) and override finalizer
+				// set large fields to null
 				disposedValue = true;
 			}
 		}
 
-		// // TODO: override finalizer only if 'Dispose(bool disposing)' has code to free unmanaged resources
+		/**
+		// // override finalizer only if 'Dispose(bool disposing)' has code to free unmanaged resources
 		// ~SingletonService()
 		// {
 		//     // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
 		//     Dispose(disposing: false);
 		// }
+		*/
 
 		public void Dispose()
 		{
@@ -103,7 +107,9 @@
 
 	public class KeyedService2 : IKeyedService;
 
+#pragma warning disable S2094 // Classes should not be empty
 	public class ExternalService;
+#pragma warning restore S2094 // Classes should not be empty
 
 	namespace Nested
 	{